change product table

+++ database/migrations/2021_11_21_152937_create_products_table.php
@@ -18,8 +18,8 @@ class CreateProductsTable extends Migration
             $table->string('name');
             $table->longText('desc')->nullable();
             $table->decimal('price', 10, 2);
-            $table->integer('tinggi')->nullable();
-            $table->integer('berat');
+            $table->decimal('tinggi', 10, 2)->nullable();
+            $table->decimal('berat', 10, 2);
             $table->string('warna')->nullable();
             $table->string('jenis')->nullable();
             $table->integer('catalog');
